split v1 and v2
